requirements-dev.txt
build
pytest
mock
pytest-cov
pytest-dotenv
pytest-flask
requests
